requirements.txt
google-api-python-client~=2.176.0
requests~=2.32.4
beautifulsoup4~=4.13.4
pandas~=2.3.1
dnspython~=2.7.0
certifi~=2025.7.9
pyOpenSSL~=25.1.0
tqdm~=4.67.1
sortedcontainers~=2.4.0